[merge] : back-develop <- S07P31D106-7/feature/signup
Squashed commit of the following:

commit fe4b2fa8ae53c07afb641b23f181dc787ef85033
Author: goodman1223 <goodman1223@naver.com>
Date:   Wed Nov 2 14:29:17 2022 +0900

    S07P31D106-7 [feat] : 중복 사용자 409 conflict 응답 기능 추가

commit 7ac740e5f6df046373fe6d529d5911e258f7cf1b
Author: goodman1223 <goodman1223@naver.com>
Date:   Wed Nov 2 14:28:59 2022 +0900

    S07P31D106-7 [feat] : validation 추가

commit fdc1ba8b7c0216d7e3a21e0dcd3032f83d1d978f
Author: goodman1223 <goodman1223@naver.com>
Date:   Wed Nov 2 14:28:40 2022 +0900

    S07P31D106-7 [fix] : imagePath 추가

commit abee1ec153f73ff64a1da0eb2aaedaeed9294a89
Author: goodman1223 <goodman1223@naver.com>
Date:   Wed Nov 2 12:55:03 2022 +0900

    [merge] : S07P31D106-7/feature/signup <- back-develop

    Squashed commit of the following:

    commit 4e4cf384cae84fc6557d38ee6c3f581b284693f0
    Author: goodman1223 <goodman1223@naver.com>
    Date:   Wed Nov 2 10:33:28 2022 +0900

        [fix] : docker file 제거

    commit 8b54bebf63cad049702776e8c883182ead526cd7
    Author: goodman1223 <goodman1223@naver.com>
    Date:   Wed Nov 2 10:33:15 2022 +0900

        [feat] : user info server uri 추가

    commit a01b84fd0c7e3570f1a3e98a11ff54ee1660c58b
    Author: goodman1223 <goodman1223@naver.com>
    Date:   Wed Nov 2 10:17:06 2022 +0900

        [fix] : 메소드 변경

    commit c3ee6594dd806ff8c705251f5110772107dce716
    Merge: 3fbd77d 1b5beed
    Author: 이상민 <goodman1223@naver.com>
    Date:   Wed Nov 2 10:06:12 2022 +0900

        Merge branch 'S07P31D106-7/feature/signup' into 'back-develop'

        [merge] :  back-develop <- S07P31D106-7/feature/signup

        See merge request s07-final/S07P31D106!40

    commit 1b5beedcef90b21f47b0610c74dfeca99f7dedaf
    Author: 이상민 <goodman1223@naver.com>
    Date:   Wed Nov 2 10:06:12 2022 +0900

        [merge] :  back-develop <- S07P31D106-7/feature/signup

commit 5eba92a0c5e891d0d57400d5cdfaae7a044b8b89
Merge: 46a75a1 3fbd77d
Author: 이상민 <goodman1223@naver.com>
Date:   Wed Nov 2 01:06:08 2022 +0000

    Merge branch 'back-develop' into 'S07P31D106-7/feature/signup'

    # Conflicts:
    #   back/iljungitjung/src/main/java/com/iljungitjung/domain/user/entity/User.java
    #   back/iljungitjung/src/main/java/com/iljungitjung/global/login/service/KakaoLoginService.java

commit 46a75a128b954c713a6511b2eceb8958f6ad0e50
Merge: 968e6a8 197911b
Author: 이상민 <goodman1223@naver.com>
Date:   Wed Nov 2 01:02:04 2022 +0000

    Merge branch 'back-develop' into 'S07P31D106-7/feature/signup'

    # Conflicts:
    #   back/iljungitjung/src/main/java/com/iljungitjung/domain/user/entity/User.java
    #   back/iljungitjung/src/main/java/com/iljungitjung/global/login/entity/RedisUser.java
    #   back/iljungitjung/src/main/java/com/iljungitjung/global/login/entity/TemporaryUser.java
    #   back/iljungitjung/src/main/java/com/iljungitjung/global/login/service/KakaoLoginService.java

+++ back/iljungitjung/src/main/java/com/iljungitjung/domain/user/service/UserServiceImpl.java
@@ -2,13 +2,13 @@ package com.iljungitjung.domain.user.service;
 
 import com.iljungitjung.domain.user.dto.SignUpDto;
 import com.iljungitjung.domain.user.entity.User;
+import com.iljungitjung.domain.user.exception.AlreadyExistUserException;
 import com.iljungitjung.domain.user.exception.NoExistUserException;
 import com.iljungitjung.domain.user.repository.UserRepository;
 import com.iljungitjung.global.login.entity.RedisUser;
 import com.iljungitjung.global.login.entity.TemporaryUser;
 import com.iljungitjung.global.login.exception.ExpireLoginUserException;
 import com.iljungitjung.global.login.exception.ExpireTemporaryUserException;
-import com.iljungitjung.global.login.exception.NotMemberException;
 import com.iljungitjung.global.login.repository.RedisUserRepository;
 import com.iljungitjung.global.login.repository.TemporaryUserRepository;
 import lombok.RequiredArgsConstructor;
@@ -48,6 +48,9 @@ public class UserServiceImpl implements UserService{
         });
         User user = signUpDto.toEntity();
         user.signUp(temporaryUser);
+        if(userRepository.existsUserByEmail(user.getEmail())){
+            throw new AlreadyExistUserException();
+        }
         log.debug("user : {}", user);
         temporaryUserRepository.deleteById(request.getSession().getId());
         userRepository.save(user);
